+++ pytest_httpx/_httpx_mock.py
@@ -1,32 +1,43 @@
 import re
-from typing import List, Union, Optional, Callable, Tuple, Pattern
+from typing import List, Union, Optional, Callable, Tuple, Pattern, Any
 
 import httpx
 import pytest
-from httpx import Request, Response, Timeout, URL, content_streams
+from httpx import Request, Response, URL, content_streams
 from httpx.dispatch.base import SyncDispatcher, AsyncDispatcher
 
 
 class _RequestMatcher:
-    def __init__(self, url: Union[str, Pattern, URL], method: str):
+    def __init__(self, url: Union[str, Pattern, URL] = None, method: str = None):
         # TODO Allow non strict URL params checking
         self.nb_calls = 0
         self.url = url
-        self.method = method.upper()
+        self.method = method
 
     def match(self, request: Request) -> bool:
         # TODO Allow to match on anything from the request
-        return self._url_match(request) and request.method == self.method
+        return self._url_match(request) and self._method_match(request)
 
     def _url_match(self, request: Request) -> bool:
+        if not self.url:
+            return True
+
         # re.Pattern was introduced in Python 3.7
         if isinstance(
             self.url, re._pattern_type if hasattr(re, "_pattern_type") else re.Pattern
         ):
             return self.url.match(str(request.url)) is not None
+
         if isinstance(self.url, str):
             return URL(self.url) == request.url
+
         return self.url == request.url
+
+    def _method_match(self, request: Request) -> bool:
+        if not self.method:
+            return True
+
+        return request.method == self.method.upper()
 
 
 class HTTPXMock:
@@ -37,18 +48,18 @@ class HTTPXMock:
 
     def add_response(
         self,
-        url: Union[str, Pattern, URL],
-        method: str = "GET",
         status_code: int = 200,
         http_version: str = "HTTP/1.1",
         headers: dict = None,
-        **content,
+        data: content_streams.RequestData = None,
+        files: content_streams.RequestFiles = None,
+        json: Any = None,
+        boundary: bytes = None,
+        **matchers,
     ):
         """
-        Mock the response that will be sent if a request is sent to this URL using this method.
+        Mock the response that will be sent if a request match.
 
-        :param url: Full URL identifying the request. Can be a str, a re.Pattern instance or a httpx.URL instance.
-        :param method: HTTP method identifying the request. Default to GET.
         :param status_code: HTTP status code of the response to send. Default to 200 (OK).
         :param http_version: HTTP protocol version of the response to send. Default to HTTP/1.1
         :param headers: HTTP headers of the response to send. Default to no headers.
@@ -57,35 +68,35 @@ class HTTPXMock:
         :param files: Multipart files.
         :param json: HTTP body of the response (if JSON should be used as content type) if data is not provided.
         :param boundary: Multipart boundary if files is provided.
+        :param url: Full URL identifying the request(s) to match. Can be a str, a re.Pattern instance or a httpx.URL instance.
+        :param method: HTTP method identifying the request(s) to match.
         """
         response = Response(
             status_code=status_code,
             http_version=http_version,
             headers=list(headers.items()) if headers else [],
-            stream=content_streams.encode(**content),
+            stream=content_streams.encode(
+                data=data, files=files, json=json, boundary=boundary
+            ),
             request=None,  # Will be set upon reception of the actual request
         )
-        self._responses.append((_RequestMatcher(url, method), response))
+        self._responses.append((_RequestMatcher(**matchers), response))
 
-    def add_callback(
-        self, callback: Callable, url: Union[str, Pattern, URL], method: str = "GET"
-    ):
+    def add_callback(self, callback: Callable, **matchers):
         """
-        Mock the action that will take place if a request is sent to this URL using this method.
+        Mock the action that will take place if a request match.
 
-        :param callback: The callable that will be called upon reception of the request.
+        :param callback: The callable that will be called upon reception of the matched request.
         It must expect at least 2 parameters:
          * request: The received request.
          * timeout: The timeout linked to the request.
         It should return an httpx.Response instance.
-        :param url: Full URL identifying the request. Can be a str, a re.Pattern instance or a httpx.URL instance.
-        :param method: HTTP method identifying the request. Default to GET.
+        :param url: Full URL identifying the request(s) to match. Can be a str, a re.Pattern instance or a httpx.URL instance.
+        :param method: HTTP method identifying the request(s) to match.
         """
-        self._callbacks.append((_RequestMatcher(url, method), callback))
+        self._callbacks.append((_RequestMatcher(**matchers), callback))
 
-    def _handle_request(
-        self, request: Request, timeout: Optional[Timeout], *args, **kwargs
-    ) -> Response:
+    def _handle_request(self, request: Request, *args, **kwargs) -> Response:
         self._requests.append(request)
 
         response = self._get_response(request)
@@ -94,7 +105,7 @@ class HTTPXMock:
 
         callback = self._get_callback(request)
         if callback:
-            return callback(request=request, timeout=timeout, *args, **kwargs)
+            return callback(request=request, *args, **kwargs)
 
         raise Exception(
             f"No mock can be found for {request.method} request on {request.url}."
@@ -146,28 +157,27 @@ class HTTPXMock:
         matcher.nb_calls += 1
         return callback
 
-    def get_requests(
-        self, url: Union[str, Pattern, URL], method: str = "GET"
-    ) -> List[Request]:
-        """
-        Return all requests sent to this URL using this method (empty list if no requests was sent).
+    # TODO Allow to assert requests content / files / whatever
 
-        :param url: Full URL identifying the requests. Can be a str, a re.Pattern instance or a httpx.URL instance.
-        :param method: HTTP method identifying the requests. Must be a upper cased string value. Default to GET.
+    def get_requests(self, **matchers) -> List[Request]:
         """
-        matcher = _RequestMatcher(url, method)
+        Return all requests sent that match (empty list if no requests were matched).
+
+        :param url: Full URL identifying the requests to retrieve. Can be a str, a re.Pattern instance or a httpx.URL instance.
+        :param method: HTTP method identifying the requests to retrieve. Must be a upper cased string value.
+        """
+        matcher = _RequestMatcher(**matchers)
         return [request for request in self._requests if matcher.match(request)]
 
-    def get_request(
-        self, url: Union[str, Pattern, URL], method: str = "GET"
-    ) -> Optional[Request]:
+    def get_request(self, **matchers) -> Optional[Request]:
         """
-        Return the request sent to this URL using this method (or None).
+        Return the single request that match (or None).
 
-        :param url: Full URL identifying the request. Can be a str, a re.Pattern instance or a httpx.URL instance.
-        :param method: HTTP method identifying the request. Must be a upper cased string value. Default to GET.
+        :param url: Full URL identifying the request to retrieve. Can be a str, a re.Pattern instance or a httpx.URL instance.
+        :param method: HTTP method identifying the request to retrieve. Must be a upper cased string value.
+        :raises AssertionError: in case more than one request match.
         """
-        requests = self.get_requests(url, method)
+        requests = self.get_requests(**matchers)
         assert (
             len(requests) <= 1
         ), f"More than one request ({len(requests)}) matched, use get_requests instead."